Remove m_CurrentInput, Enforce 1 input per tick
This converts m_aInputs from an unordered buffer of inputs that can be intended for any tick into a strict ring buffer that tracks the 100 inputs ahead of the current gametick and 100 inputs behind.

The client is only allowed to have 1 input for each gametick in the buffer and cannot change inputs for ticks they have already sent.